+++ wazo_auth/plugins/http/tokens/plugin.py
@@ -17,6 +17,7 @@ class Plugin:
         api.add_resource(
             http.Token, '/token/<string:token_uuid>', resource_class_args=args
         )
+        api.add_resource(http.RefreshTokens, '/tokens', resource_class_args=args)
         api.add_resource(
             http.UserRefreshTokens,
             '/users/<uuid:user_uuid>/tokens',