+++ wazo_auth/plugins/http/tokens/schemas.py
@@ -59,3 +59,5 @@ class RefreshTokenSchema(BaseSchema):
     client_id = xfields.String(min=1, max=1024)
     created_at = xfields.DateTime()
     mobile = xfields.Boolean()
+    user_uuid = xfields.String()
+    tenant_uuid = xfields.String()